cec: fix a standby issue when cec auto device off is disabled [1/1]
PD#SWPL-18636

Problem:
The auto device off is false by default set in HdmiControlService.
But the value got in hal is true by default, it's not good.If you
don't change the options in settigs, the problem keeps staying.
This issue is introduced by
http://scgit.amlogic.com:8080/#/c/76458

Solution:
Temporarily change the default value in hal to false. I will
refactor the settings logic in SWPL-19011.

Verify:
newton

Change-Id: I040bd056ea44debf2d5a97e17b8e78d90aa64d2a
Signed-off-by: an.xi <an.xi@amlogic.com>