dolbyvision: need check deep color when check the dv state. [1/1]
PD#SWPL-19664

Problem:
systemcontrol check the dolby_vision_ll_policy and dolby_vision_enable.
but not check the deep color.

Solution:
check deep color when check the dv state.
if dv state is STD but the attr is not 444,8bit
or dv state is LL but the attr is not 422,12bit
or dv state is LL-RGB but the attr is not 444,10bit/444,12bit
need set dv state again.

Verify:
verify by newton.

Change-Id: I2143d7a4e163a02342fc4a2768976fd7bade658c
Signed-off-by: Luan Yuan <luan.yuan@amlogic.com>